refactor: move `DashboardSummary` type definition to `src/types` for improved modularity
- Updated all relevant imports across components and services.
- Removed redundant definition in `dashboard-service`.
