+++ src/components/dashboard/TotalSignedEventsCard.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
 import { useI18n } from 'vue-i18n';
-import { type DashboardSummary } from 'src/services/dashboard-service';
+import { type DashboardSummary } from 'src/types';
 
 withDefaults(
   defineProps<{
@@ -14,8 +14,6 @@ withDefaults(
     error: null,
   },
 );
-
-
 const { t } = useI18n();
 </script>
