+++ src/types/index.d.ts
@@ -1,6 +1,11 @@
 import type { Event as NostrEvent, UnsignedEvent } from 'nostr-tools';
 import type { ErrorCode } from 'src/types/error-codes';
 import type { QuickSignSupportedKind, QuickSignTagType } from 'src/services/quick-sign-service';
+import type {
+  ConnectedRelaysDataState,
+  DashboardActivityItem,
+  DashboardDataState,
+} from 'src/services/dashboard-service';
 
 export interface DropdownItem<T = string | number> {
   label: string;
@@ -124,4 +129,13 @@ interface DataActivityRow {
   time: string;
   statusLabel: string;
   statusVariant: ActivityStatusVariant;
+}
+
+export interface DashboardSummary {
+  state: DashboardDataState;
+  signedEvents: number;
+  activeKeys: number;
+  connectedRelays: number;
+  connectedRelaysState: ConnectedRelaysDataState;
+  recentActivity: DashboardActivityItem[];
 }
